[~] Added logic to method 'GetISBN', change properties 'ISBN' value in 'Book' class from int to long, added new migration to database

+++ Library/Migrations/LibraryDbContextModelSnapshot.cs
@@ -37,8 +37,8 @@ namespace Library.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("ISBN")
-                        .HasColumnType("int");
+                    b.Property<long>("ISBN")
+                        .HasColumnType("bigint");
 
                     b.Property<bool>("IsAvailable")
                         .HasColumnType("bit");
